program-runtime: port over process_message from svm
