+++ svm/src/conformance/instr/harness.rs
@@ -2,15 +2,12 @@
 
 use {
     super::{context::InstrContext, effects::InstrEffects},
-    crate::{
-        conformance::{
-            callback::DefaultCallback,
-            setup::{
-                InvokeContextFields, compute_budget, prepare_invoke_context_fields,
-                program_runtime_environments,
-            },
+    crate::conformance::{
+        callback::DefaultCallback,
+        setup::{
+            InvokeContextFields, compute_budget, prepare_invoke_context_fields,
+            program_runtime_environments,
         },
-        message_processor::process_message,
     },
     solana_instruction::error::InstructionError,
     solana_program_runtime::{
@@ -92,9 +89,8 @@ pub fn execute_instr_with_callback<C: InvokeContextCallback>(
             execution_cost,
         );
 
-        match process_message(
+        match invoke_context.process_message(
             &sanitized_message,
-            &mut invoke_context,
             &mut timings,
             &mut compute_units_consumed,
         ) {
